PUMICE.compute_weight {PUMICE}	R Documentation

Create gene expression prediction model using window type and penalty factor derived from the CV step

Description

Create gene expression prediction model using window type and penalty factor derived from the CV step

Usage

PUMICE.compute_weight(
  geno_path,
  exp_path,
  encode_path,
  output_path,
  window_path_list
)

Arguments

geno_path
genotype path, plink format

exp_path
expression path, txt format

encode_path
epigenomic annotation path, txt format

output_path
dir of your output file

window_path_list
A list of 3d genomic winodows you need

Examples

PUMICE.compute_weight()

[Package PUMICE version 0.0.0.9000 Index]