PUMICE.nested_cv {PUMICE}	R Documentation

Run nested cross-validation to determine which window type and penalty factor are optimal

Description

Run nested cross-validation to determine which window type and penalty factor are optimal

Usage

PUMICE.nested_cv(
  geno_path,
  exp_path,
  encode_path,
  output_path,
  window,
  window_path_list = NULL
)

Arguments

geno_path
genotype path, plink format

exp_path
expression path, txt format

encode_path
epigenomic annotation path, txt format

output_path
dir of your output file

window
specify the constant or 3D window you need

window_path_list
A list of 3d genomic winodows you need

Examples

PUMICE.nested_cv()

[Package PUMICE version 0.0.0.9000 Index]