expression_scale {PUMICE}	R Documentation

Function for scaling expression dataframe to mean of zero

Description

Function for scaling expression dataframe to mean of zero

Usage

expression_scale(expression_subset, sample_tuning, sample_testing)

Arguments

expression_subset
input expression

sample_tuning
index of tunning sample

sample_testing
index of testing sample

Examples

expression_scale()

[Package PUMICE version 0.0.0.9000 Index]