geno_scale {PUMICE}	R Documentation

Function for scaling genotype dataframe to mean of zero and sd of one

Description

Function for scaling genotype dataframe to mean of zero and sd of one

Usage

geno_scale(geno_subset, sample_tuning, sample_testing)

Arguments

geno_subset
input genotype matrix

sample_tuning
index of tunning sample

sample_testing
index of testing sample

Examples

geno_scale()

[Package PUMICE version 0.0.0.9000 Index]